Initial Commit
Able to create random background composed of random rectangles and arcs with varying colours and transparencies. Animation interval known to bring Firefox to a halt when set below 500ms. Tested as working as low as 100ms on Chrome, Chrome-based Edge (on Windows 10), and iOS Safari. Looking to add more descriptive console messages and reform the UI for controls.